Refactored teams to smaller components for TeamCard, TeamHeader, TeamList and centralised the interfaces
